fix non-merged test build
